Move game branch creation to separate subcommand
